refactor map widget
